fix(handlers): 移除 count_tokens 的多余 i32 cast(存量 clippy)
total_tokens 已是 i32，.max(1) as i32 的 cast 冗余。修掉让 #44 CI lint 转绿。

#43

+++ src/anthropic/handlers.rs
@@ -903,7 +903,7 @@ pub async fn count_tokens(
     ) as i32;
 
     Json(CountTokensResponse {
-        input_tokens: total_tokens.max(1) as i32,
+        input_tokens: total_tokens.max(1),
     })
 }
